+++ lib/bento/runner.rb
@@ -37,7 +37,7 @@ class BuildRunner
     templates = template_files
     banner('Starting build for templates:')
     banner('Installing packer plugins') unless dry_run || metadata_only
-    shellout("packer init -upgrade #{File.dirname(templates.first)}/../../packer_templates") unless dry_run || metadata_only
+    shellout("packer init -upgrade #{File.absolute_path("#{File.dirname(templates.first)}/../../packer_templates")}") unless dry_run || metadata_only
     templates.each { |t| puts "- #{t}" }
     time = Benchmark.measure do
       templates.each { |template| build(template) }
@@ -81,7 +81,7 @@ class BuildRunner
 
   def packer_build_cmd(template, _var_file)
     pkrvars = "#{template}.pkrvars.hcl"
-    cmd = %W(packer build -timestamp-ui -force -var-file=#{pkrvars} ../../packer_templates)
+    cmd = %W(packer build -timestamp-ui -force -var-file=#{File.absolute_path(pkrvars)} #{File.absolute_path("../../packer_templates")})
     vars.each do |var|
       cmd.insert(4, "-var #{var}")
     end if vars